:recycle: refactor(bid): Move to Bid domain, Unify controller versioning
