+++ src/main/java/nbc/mushroom/domain/bid/controller/BidController.java
@@ -1,41 +1,64 @@
 package nbc.mushroom.domain.bid.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import nbc.mushroom.domain.bid.dto.request.CreateBidReq;
-import nbc.mushroom.domain.bid.dto.response.CreateBidRes;
+import nbc.mushroom.domain.bid.dto.response.BidRes;
 import nbc.mushroom.domain.bid.service.BidService;
 import nbc.mushroom.domain.common.annotation.Auth;
 import nbc.mushroom.domain.common.dto.ApiResponse;
 import nbc.mushroom.domain.common.dto.AuthUser;
 import nbc.mushroom.domain.user.entity.User;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/api/users/bids")
 @RequiredArgsConstructor
-@RequestMapping("/api/auction-items/{auctionItemId}/bids")
 public class BidController {
 
     private final BidService bidService;
 
-    @PostMapping
-    public ResponseEntity<ApiResponse<CreateBidRes>> createOrUpdateBid(
+    @GetMapping
+    public ResponseEntity<ApiResponse<Page<BidRes>>> getAllBids(
         @Auth AuthUser authUser,
-        @PathVariable Long auctionItemId,
-        @Valid @RequestBody CreateBidReq createBidReq
+        @RequestParam(defaultValue = "1") int page
     ) {
-        CreateBidRes createBidRes = bidService.createOrUpdateBid(
-            User.fromAuthUser(authUser),
-            auctionItemId,
-            createBidReq);
+        Pageable pageable = PageRequest.of(page - 1, 10);
+        User loginUser = User.fromAuthUser(authUser);
 
-        return ResponseEntity.status(HttpStatus.CREATED)
-            .body(ApiResponse.success("입찰에 성공했습니다.", createBidRes));
+        Page<BidRes> getUserBidRes = bidService.getAllBidsByUser(loginUser, pageable);
+
+        return ResponseEntity.status(HttpStatus.OK)
+            .body(ApiResponse.success("입찰 내역 목록 조회에 성공했습니다", getUserBidRes));
+    }
+
+    @GetMapping("/{bidId}")
+    public ResponseEntity<ApiResponse<BidRes>> getBid(
+        @Auth AuthUser authUser,
+        @PathVariable Long bidId
+    ) {
+        User loginUser = User.fromAuthUser(authUser);
+        BidRes bidRes = bidService.getBidByUser(loginUser, bidId);
+
+        return ResponseEntity.status(HttpStatus.OK)
+            .body(ApiResponse.success("입찰 내역 상세 조회에 성공했습니다.", bidRes));
+    }
+
+    @DeleteMapping("/{bidId}/cancel")
+    public ResponseEntity<ApiResponse<Void>> deleteBid(
+        @Auth AuthUser authUser,
+        @PathVariable Long bidId
+    ) {
+        bidService.deleteBid(User.fromAuthUser(authUser), bidId);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT)
+            .body(ApiResponse.success("입찰 취소에 성공했습니다."));
     }
 }
