+++ src/main/java/nbc/mushroom/domain/bid/dto/response/BidRes.java
@@ -1,18 +1,18 @@
-package nbc.mushroom.domain.user.dto.response;
+package nbc.mushroom.domain.bid.dto.response;
 
 import nbc.mushroom.domain.auction_item.dto.response.SearchAuctionItemRes;
 import nbc.mushroom.domain.bid.entity.Bid;
 import nbc.mushroom.domain.bid.entity.BiddingStatus;
 
-public record UserBidRes(
+public record BidRes(
     Long bidId,
     Long biddingPrice,
     BiddingStatus biddingStatus,
     SearchAuctionItemRes searchAuctionItemRes
 ) {
 
-    public static UserBidRes from(Bid bid, SearchAuctionItemRes searchAuctionItemRes) {
-        return new UserBidRes(
+    public static BidRes from(Bid bid, SearchAuctionItemRes searchAuctionItemRes) {
+        return new BidRes(
             bid.getId(),
             bid.getBiddingPrice(),
             bid.getBiddingStatus(),
